// frontend/src/api/generated.d.ts
export interface components {
    schemas: {
// ...
        CreateReconciliation: {
            /** Format: uuid */
            financialAccountId: string;
            /** Format: date */
            statementDate: string;
            statementBalance: components["schemas"]["MoneyAmount"];
            complete: boolean;
        };
// ...
    };
// ...
}